moving event files to /events
