+++ packages/driver/src/multi-domain/cypress.ts
@@ -12,12 +12,12 @@ import { bindToListeners } from '../cy/listeners'
 import { SpecBridgeDomainCommunicator } from './communicator'
 import { handleDomainFn } from './domain_fn'
 import { handleCommands } from './commands'
-import { handleLogs } from './logs'
-import { handleSocketEvents } from './socket'
-import { handleSpecWindowEvents } from './spec_window_events'
-import { handleErrorEvent } from './errors'
-import { handleScreenshots } from './screenshots'
-import { handleTestEvents } from './test_events'
+import { handleLogs } from './events/logs'
+import { handleSocketEvents } from './events/socket'
+import { handleSpecWindowEvents } from './events/spec_window_events'
+import { handleErrorEvent } from './events/errors'
+import { handleScreenshots } from './events/screenshots'
+import { handleTestEvents } from './events/test_events'
 
 const specBridgeCommunicator = new SpecBridgeDomainCommunicator()
